requirements.txt
langchain
langchain-openai
langchain_community
streamlit
numpy
python-dotenv
transformers 
torch 
TextBlob 
matplotlib 
plotly 
psycopg2-binary

